Deleted some useless files and renamed some directories.
