+++ .vscode/settings.json
@@ -5,7 +5,6 @@
   "spellright.documentTypes": [
     "markdown",
     "latex",
-    "plaintext",
-    "javascript"
+    "plaintext"
   ]
 }
