--- src/Test_Objects/jsonApi.js
@@ -1,3102 +1,0 @@
-const api = {
-  "Alabama":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Alaska":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Arizona":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Arkansas":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "California":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Colorado":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Connecticut":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Delaware":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Florida":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Georgia":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Hawaii":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Idaho":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Illinois":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Indiana":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Iowa":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Kansas":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Kentucky":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Louisiana":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Maine":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Maryland":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Massachusetts":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Michigan":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Minnesota":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Mississippi":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Missouri":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Montana":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Nebraska":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Nevada":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "New Hampshire":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "New Jersey":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "New Mexico":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "New York":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "North Carolina":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "North Dakota":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Ohio":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Oklahoma":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Oregon":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Pennsylvania":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Rhode Island":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "South Carolina":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "South Dakota":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Tennessee":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Texas":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Utah":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Vermont":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Virginia":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Washington":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "West Virginia":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Wisconsin":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  },
-  "Wyoming":{
-     "January":{
-        "produce":[
-
-        ]
-     },
-     "February":{
-        "produce":[
-
-        ]
-     },
-     "March":{
-        "produce":[
-
-        ]
-     },
-     "April":{
-        "produce":[
-
-        ]
-     },
-     "May":{
-        "produce":[
-
-        ]
-     },
-     "June":{
-        "produce":[
-
-        ]
-     },
-     "July":{
-        "produce":[
-
-        ]
-     },
-     "August":{
-        "produce":[
-
-        ]
-     },
-     "September":{
-        "produce":[
-
-        ]
-     },
-     "October":{
-        "produce":[
-
-        ]
-     },
-     "November":{
-        "produce":[
-
-        ]
-     },
-     "December":{
-        "produce":[
-
-        ]
-     }
-  }
-}
